backend/requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Validation
pydantic==2.5.3
email-validator==2.1.0

# Géospatial (pour PostGIS)
geoalchemy2==0.14.3
shapely==2.0.2

# Rapports et exports
reportlab==4.0.9
openpyxl==3.1.2
pandas==2.1.4

# Utilitaires
python-dateutil==2.8.2

# Tests
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# CORS
python-cors==1.0.0